# app/session7-deployment/backend/api.yml
openapi: 3.0.3
info:
  title: Mini EASM API - Session 6 (Testing)
  description: |
    # External Attack Surface Management API

    API hoàn chỉnh cho hệ thống EASM với các tính năng:
    - ✅ Asset Management (CRUD operations)
    - ✅ Scanning Operations (DNS, WHOIS, Subdomain)
    - ✅ Scan Job Management
    - ✅ Result Retrieval & Filtering
    - ✅ Pagination Support
    - 🧪 Full Unit Test Coverage (26 tests + 6 benchmarks)

    ## Sessions Progression:
    - Session 2: Basic API (CRUD)
    - Session 3: Database Integration (PostgreSQL)
    - Session 4: Advanced API (Filtering, Pagination, Validation)
    - Session 5: EASM Scanning (DNS, WHOIS, Subdomain Enumeration)
    - Session 6: Testing & Quality Assurance ← Current

    ## Scan Types:

    ### 🟢 Passive Scans (Safe - No Permission Required)
    - `dns` - DNS record queries (A, AAAA, MX, NS, TXT, CNAME)
    - `whois` - WHOIS registration information
    - `subdomain` - Subdomain enumeration via DNS bruteforce
    - `cert_trans` - Certificate Transparency logs
    - `asn` - ASN lookup
    - `all` - Run all passive scans

    ### 🔴 Active Scans (⚠️ Requires Permission)
    - `port` - Port scanning (requires authorization)
    - `ssl` - SSL/TLS certificate probing (requires authorization)

    **⚠️ LEGAL WARNING:** Active scans may be illegal without explicit permission from target owner.
    Only scan systems you own or have written authorization to test.

  version: 6.0.0
  contact:
    name: CMC Intern Program
...
servers:
...
paths:
...
  /assets/{id}/scan:
    post:
      summary: Start a scan job (async)
      description: |
        Khởi động một scan job cho asset. Scan chạy asynchronously trong background.

        **Workflow:**
        1. POST /assets/{id}/scan - Start scan (returns job_id)
        2. GET /scan-jobs/{job_id} - Poll for status
        3. GET /scan-jobs/{job_id}/results - Get results when completed

        **Scan Types:**
        - 🟢 `dns` - DNS record lookup (passive)
        - 🟢 `whois` - WHOIS information (passive)
        - 🟢 `subdomain` - Subdomain enumeration (passive)
        - 🟢 `cert_trans` - Certificate Transparency (passive)
        - 🟢 `asn` - ASN lookup (passive)
        - 🟢 `all` - All passive scans
        - 🔴 `port` - Port scanning (⚠️ requires permission)
        - 🔴 `ssl` - SSL/TLS probing (⚠️ requires permission)

      operationId: startScan
      tags:
        - Scanning
      parameters:
        - $ref: "#/components/parameters/AssetId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - scan_type
              properties:
                scan_type:
                  $ref: "#/components/schemas/ScanType"
            examples:
              dns_scan:
                summary: DNS scan
                value:
                  scan_type: dns
              ip_scan:
                summary: IP scan
                value:
                  scan_type: ip
              port_scan:
                summary: Port scan
                value:
                  scan_type: port
              whois_scan:
                summary: WHOIS scan
                value:
                  scan_type: whois
              subdomain_scan:
                summary: Subdomain enumeration
                value:
                  scan_type: subdomain
              all_passive:
                summary: All passive scans
                value:
                  scan_type: all
      responses:
        "202":
          description: Scan job accepted (async processing started)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScanJob"
        "400":
          description: Invalid request (invalid scan type or asset ID)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Asset not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /assets/{id}/scan/demo:
    post:
      summary: Demo endpoint - Sync vs Async comparison
      description: |
        🎓 **Educational endpoint** for demonstrating sync vs async patterns.

        Runs the same scans twice:
        1. Synchronously (blocking)
        2. Asynchronously (non-blocking)

        Returns timing comparison to show performance difference.

        **Use case:** Learning about async programming patterns in Go.

      operationId: demoSyncVsAsync
      tags:
        - Scanning
      parameters:
        - $ref: "#/components/parameters/AssetId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                scan_types:
                  type: array
                  items:
                    $ref: "#/components/schemas/ScanType"
                  example: [dns, whois]
            examples:
              demo:
                summary: Compare DNS and WHOIS
                value:
                  scan_types: [dns, whois]
      responses:
        "200":
          description: Demo completed with timing comparison
          content:
            application/json:
              schema:
                type: object
                properties:
                  sync_duration_ms:
                    type: number
                    example: 2500
                  async_duration_ms:
                    type: number
                    example: 1200
                  improvement:
                    type: string
                    example: "52% faster"
                  sync_jobs:
                    type: array
                    items:
                      $ref: "#/components/schemas/ScanJob"
                  async_jobs:
                    type: array
                    items:
                      $ref: "#/components/schemas/ScanJob"
        "404":
          description: Asset not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /assets/{id}/scans:
    get:
      summary: List all scan jobs for an asset
      description: |
        Lấy danh sách tất cả scan jobs của một asset với filtering và pagination.

        **Filters:**
        - `scan_type`: Filter by scan type
        - `status`: Filter by status
        - `page`: Page number
        - `page_size`: Items per page

      operationId: listScanJobs
      tags:
        - Scanning
      parameters:
        - $ref: "#/components/parameters/AssetId"
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: page_size
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: scan_type
          in: query
          description: Filter by scan type
          schema:
            $ref: "#/components/schemas/ScanType"
        - name: status
          in: query
          description: Filter by status
          schema:
            $ref: "#/components/schemas/ScanStatus"
      responses:
        "200":
          description: Paginated list of scan jobs
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedScanJobs"
        "404":
          description: Asset not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /scan-jobs/{id}:
    get:
      summary: Get scan job status
      description: |
        Lấy thông tin chi tiết và status của một scan job.

        **Status Flow:**
        1. `pending` - Job queued but not started
        2. `running` - Job is executing
        3. `completed` - Job finished successfully
        4. `failed` - Job encountered error
        5. `partial` - Job completed with some errors

      operationId: getScanJob
      tags:
        - Scanning
      parameters:
        - $ref: "#/components/parameters/ScanJobId"
      responses:
        "200":
          description: Scan job found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScanJob"
        "404":
          description: Scan job not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /scan-jobs/{id}/results:
    get:
      summary: Get scan job results
      description: |
        Lấy tất cả results của một scan job.

        **Result types based on scan_type:**
        - `dns` → DNSRecord[]
        - `whois` → WHOISRecord
        - `subdomain` → Subdomain[]
        - `ip` → IPScanResult[]
        - `port` → PortScanResult[]
        - `all` → Mixed results

      operationId: getScanResults
      tags:
        - Results
      parameters:
        - $ref: "#/components/parameters/ScanJobId"
      responses:
        "200":
          description: Scan results retrieved
          content:
            application/json:
              schema:
                type: object
                properties:
                  job_id:
                    type: string
                    format: uuid
                  scan_type:
                    $ref: "#/components/schemas/ScanType"
                  results:
                    type: array
                    items:
                      oneOf:
                        - $ref: "#/components/schemas/DNSRecord"
                        - $ref: "#/components/schemas/WHOISRecord"
                        - $ref: "#/components/schemas/Subdomain"
                        - $ref: "#/components/schemas/IPScanResult"
                        - $ref: "#/components/schemas/PortScanResult"
        "404":
          description: Scan job not found or no results
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  # ============================================
  # RESULT RETRIEVAL (Aggregated by Asset)
  # ============================================
  /assets/{id}/results:
    get:
      summary: Get ALL scan results for an asset
      description: |
        Lấy tất cả scan results của một asset (DNS + WHOIS + Subdomains).
        Results từ tất cả scan jobs đã chạy.

      operationId: getAssetAllResults
      tags:
        - Results
      parameters:
        - $ref: "#/components/parameters/AssetId"
      responses:
        "200":
          description: All results retrieved
          content:
            application/json:
              schema:
                type: object
                properties:
                  asset_id:
                    type: string
                    format: uuid
                  dns_records:
                    type: array
                    items:
                      $ref: "#/components/schemas/DNSRecord"
                  whois_records:
                    type: array
                    items:
                      $ref: "#/components/schemas/WHOISRecord"
                  subdomains:
                    type: array
                    items:
                      $ref: "#/components/schemas/Subdomain"
        "404":
          description: Asset not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /assets/{id}/subdomains:
    get:
      summary: Get subdomains for an asset
      description: |
        Lấy danh sách tất cả subdomains đã discovered cho asset.
        Supports pagination.

      operationId: getAssetSubdomains
      tags:
        - Results
      parameters:
        - $ref: "#/components/parameters/AssetId"
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PageSize"
      responses:
        "200":
          description: Subdomains retrieved
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/Subdomain"
                  total:
                    type: integer
                  page:
                    type: integer
                  page_size:
                    type: integer
                  total_pages:
                    type: integer
        "404":
          description: Asset not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /assets/{id}/dns:
    get:
      summary: Get DNS records for an asset
      description: |
        Lấy danh sách tất cả DNS records cho asset.

        **Optional filter:**
        - `record_type`: Filter by DNS record type (A, AAAA, MX, NS, TXT, CNAME, SOA)

      operationId: getAssetDNS
      tags:
        - Results
      parameters:
        - $ref: "#/components/parameters/AssetId"
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PageSize"
        - name: record_type
          in: query
          description: Filter by DNS record type
          schema:
            type: string
            enum: [A, AAAA, MX, NS, TXT, CNAME, SOA]
            example: A
      responses:
        "200":
          description: DNS records retrieved
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/DNSRecord"
                  total:
                    type: integer
                  page:
                    type: integer
                  page_size:
                    type: integer
                  total_pages:
                    type: integer
        "404":
          description: Asset not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /assets/{id}/whois:
    get:
      summary: Get WHOIS record for an asset
      description: |
        Lấy WHOIS information cho asset.
        Returns single most recent WHOIS record.

      operationId: getAssetWHOIS
      tags:
        - Results
      parameters:
        - $ref: "#/components/parameters/AssetId"
      responses:
        "200":
          description: WHOIS record retrieved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WHOISRecord"
        "404":
          description: Asset not found or no WHOIS data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

# ============================================
# REUSABLE COMPONENTS
# ============================================
components:
  parameters:
    AssetId:
      name: id
      in: path
      required: true
      description: Asset UUID
      schema:
        type: string
        format: uuid
        example: "123e4567-e89b-12d3-a456-426614174000"
    ScanJobId:
      name: id
      in: path
      required: true
      description: Scan job UUID
      schema:
        type: string
        format: uuid
        example: "987e6543-e21b-34c5-d678-426614174999"
    Page:
      name: page
      in: query
      description: Page number (starts from 1)
      schema:
        type: integer
        minimum: 1
        default: 1
    PageSize:
      name: page_size
      in: query
      description: Items per page
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
  schemas:
...
    ScanJob:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: "987e6543-e21b-34c5-d678-426614174999"
        asset_id:
          type: string
          format: uuid
          example: "123e4567-e89b-12d3-a456-426614174000"
        scan_type:
          $ref: "#/components/schemas/ScanType"
        status:
          $ref: "#/components/schemas/ScanStatus"
        started_at:
          type: string
          format: date-time
          example: "2026-03-09T10:00:00Z"
        ended_at:
          type: string
          format: date-time
          nullable: true
          example: "2026-03-09T10:00:15Z"
        error:
          type: string
          example: ""
          description: Error message if failed
        results:
          type: integer
          example: 25
          description: Number of results found
        created_at:
          type: string
          format: date-time
          example: "2026-03-09T10:00:00Z"
    DNSRecord:
      type: object
      properties:
        id:
          type: string
          format: uuid
        asset_id:
          type: string
          format: uuid
        scan_job_id:
          type: string
          format: uuid
        record_type:
          type: string
          enum: [A, AAAA, MX, NS, TXT, CNAME, SOA]
          example: A
        name:
          type: string
          example: example.com
        value:
          type: string
          example: 93.184.216.34
        ttl:
          type: integer
          example: 3600
          description: Time to live in seconds
        created_at:
          type: string
          format: date-time
    WHOISRecord:
      type: object
      properties:
        id:
          type: string
          format: uuid
        asset_id:
          type: string
          format: uuid
        scan_job_id:
          type: string
          format: uuid
        registrar:
          type: string
          example: "MarkMonitor Inc."
        created_date:
          type: string
          format: date-time
          nullable: true
          example: "1995-08-14T00:00:00Z"
        expiry_date:
          type: string
          format: date-time
          nullable: true
          example: "2027-08-13T00:00:00Z"
        name_servers:
          type: string
          example: '["ns1.example.com", "ns2.example.com"]'
          description: JSON array of nameservers
        status:
          type: string
          example: "clientTransferProhibited"
        emails:
          type: string
          example: '["admin@example.com"]'
          description: JSON array of contact emails
        raw_data:
          type: string
          example: "Domain Name: EXAMPLE.COM\n..."
          description: Full raw WHOIS response
        created_at:
          type: string
          format: date-time
    Subdomain:
      type: object
      properties:
        id:
          type: string
          format: uuid
        asset_id:
          type: string
          format: uuid
          description: Parent domain asset
        scan_job_id:
          type: string
          format: uuid
        name:
          type: string
          example: api.example.com
        source:
          type: string
          example: dns_bruteforce
          description: Discovery method
        is_active:
          type: boolean
          example: true
          description: Whether subdomain is reachable
        created_at:
          type: string
          format: date-time
    IPScanResult:
      type: object
      properties:
        id:
          type: string
          format: uuid
        asset_id:
          type: string
          format: uuid
        scan_job_id:
          type: string
          format: uuid
        ip_address:
          type: string
          example: 127.0.0.1
        geolocation:
          type: object
          properties:
            country:
              type: string
              example: Local
            country_code:
              type: string
              example: LO
            city:
              type: string
              example: Localhost
            region:
              type: string
              example: Loopback
            latitude:
              type: number
              format: double
              example: 0
            longitude:
              type: number
              format: double
              example: 0
            isp:
              type: string
              example: Local Network
            org:
              type: string
              example: Training
        asn:
          type: object
          properties:
            number:
              type: integer
              example: 0
            name:
              type: string
              example: LOCAL
            description:
              type: string
              example: Private/local network
        reverse_dns:
          type: string
          example: localhost
        created_at:
          type: string
          format: date-time
...
    PortScanResult:
      type: object
      properties:
        id:
          type: string
          format: uuid
        asset_id:
          type: string
          format: uuid
        scan_job_id:
          type: string
          format: uuid
        ip_address:
          type: string
          example: 127.0.0.1
        open_ports:
          type: array
          items:
            $ref: "#/components/schemas/OpenPort"
        closed_ports:
          type: integer
          example: 12
        total_scanned:
          type: integer
          example: 16
        scan_duration_ms:
          type: integer
          example: 250
        created_at:
          type: string
          format: date-time

    # ============================================
    # ENUMS
    # ============================================
    ScanType:
      type: string
      enum:
        - all
        - dns
        - whois
        - subdomain
        - cert_trans
        - asn
        - ip
        - port
        - ssl
      example: dns
      description: |
        Scan types:
        - `all`: Run all passive scans
        - `dns`: DNS record lookup (passive)
        - `whois`: WHOIS information (passive)
        - `subdomain`: Subdomain enumeration (passive)
        - `cert_trans`: Certificate Transparency logs (passive)
        - `asn`: ASN lookup (passive)
        - `ip`: IP geolocation + ASN + reverse DNS (passive)
        - `port`: Port scanning (⚠️ active - requires permission)
        - `ssl`: SSL/TLS probing (⚠️ active - requires permission)
    ScanStatus:
      type: string
      enum:
        - pending
        - running
        - completed
        - failed
        - partial
      example: completed
      description: |
        Scan status:
        - `pending`: Job queued but not started
        - `running`: Job is executing
        - `completed`: Job finished successfully
        - `failed`: Job encountered fatal error
        - `partial`: Job completed with some errors

    # ============================================
    # PAGINATION
    # ============================================
    PaginatedScanJobs:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/ScanJob"
        total:
          type: integer
          example: 150
        page:
          type: integer
          example: 1
        page_size:
          type: integer
          example: 20
        total_pages:
          type: integer
          example: 8

    # ============================================
    # ERRORS
    # ============================================
    Error:
      type: object
      required:
        - error
      properties:
        error:
          type: string
          example: "invalid request body"
          description: Error message describing what went wrong